+++ src/Servers/OSX.php
@@ -24,7 +24,7 @@ final class OSX implements Implementation
     private MemoryFacade $memory;
     private Processes\Unix $processes;
     private LoadAverageFacade $loadAverage;
-    private Disk\Unix $disk;
+    private Disk $disk;
 
     private function __construct(Clock $clock, Control $control, EnvironmentPath $path)
     {
@@ -32,7 +32,7 @@ final class OSX implements Implementation
         $this->memory = new MemoryFacade($control->processes(), $path);
         $this->processes = Processes\Unix::osx($clock, $control->processes());
         $this->loadAverage = new LoadAverageFacade;
-        $this->disk = Disk\Unix::of($control->processes());
+        $this->disk = Disk::of($control->processes());
     }
 
     /**