+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Innmind\Server\Status\Server` is now a final class, all previous implementations are now flagged as internal
 - `Innmind\Server\Status\Server\Processes::all()` now returns an `Innmind\Immutable\Sequence`
 - `Innmind\Server\Status\Server\Disk::volumes()` now returns an `Innmind\Immutable\Sequence`
+- `Innmind\Server\Status\Server\Disk` is now a final class
 
 ## 5.0.0 - 2025-05-08